+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-22T11:53:07.927Z",
+    "2026-07-22T14:00:53.042Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {